fix: 연속 일정 버그 수정

+++ app/src/main/java/com/example/haru/view/calendar/CalendarAddFragment.kt
@@ -564,7 +564,7 @@ class CalendarAddFragment(private val activity: Activity,
                         binding.repeatStartTimeBtn.text.toString().substring(6,8).toInt()
                     )
 
-                    repeatvalue = "T"+(repeatEndCalendar.time.time - repeatStartCalendar.time.time).toString()
+                    repeatvalue = "T"+(repeatEndCalendar.time.time - repeatStartCalendar.time.time).toInt().toString()
                 }
             }
